requirements.txt
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
python-pptx==0.6.21
PyPDF2==3.0.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
numpy<2.0.0
huggingface-hub==0.19.4 